requirements.txt
cryptography
langchain
google-generativeai
python-dotenv
